Add configmap support to the accelerator's collector
accelerator collector needs to have a configuration that describes which
vendors and models it should collect. This PR add loading that data from
a file. Eventually this file be a configmap mapped into the
node-exporter pod.
Parsing the configmap file is done in the init function. The file is
defined by an environment variable. This variable will be configured by
CMO and by unit-test