Add configmap support to the accelerator's collector
accelerator collector needs to have a configuration that describes which
vendors and models it should collect. This PR adds loading that data from
a file. This file is actually a configmap mapped into the
node-exporter pod by cluster-monitoring-operator.
The file location is defined by the argument passed to node-exporter
executable, or by its default value, if no argument is passed.